Update song name on screen  via MQTT
This commit updates the showPlayerScene function to check for new song names received via mqtt and display the song name on the screen. The function now includes an if statement that checks the messageReceived boolean flag and updates the song name on the display if a new message has been received. The updateSongName function has also been updated to set the messageReceived flag when a new song name is received. In addition a new mqtt topic for song names has been created to send song names via mqtt